+++ changelog.md
@@ -94,3 +94,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 - Added fallback and catch exeptions for teleportation locations
 
+## [0.2.35] - 2024-10-26
+
+### Added
+- version in the c# code, it will log the version in the logs
+
+### Fixed
+- Bug in construct panel when the construct was owned by a organization
+- Bug that would spawn items in linked container instead of nanopack
+